OCR_Python/requirements_adj.txt
boto3>=1.33.2
botocore>=1.33.4
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.2
contourpy>=1.2.0
cryptography>=41.0.7
cycler>=0.12.1
fonttools>=4.45.1
google-cloud>=0.34.0
jmespath>=1.0.1
kiwisolver>=1.4.5
matplotlib>=3.8.2
numpy>=1.26.2
opencv-python>=4.8.1
packaging>=23.2
pandas>=2.1.3
pdfminer.six>=20221105
Pillow>=10.1.0
pluginbase>=1.0.1
pycparser>=2.21
pyparsing>=3.1.1
python-dateutil>=2.8.2
pytz>=2023.3
s3transfer>=0.8.1
six>=1.16.0
TExtractor>=0.1.2
tzdata>=2023.3
urllib3>=1.25.4